Fix memory error in float(str) conversion
We need to check for success in `digitsFromDigitsWithUnderscores`, not
if `new_str == nullptr`. The latter is extremely unlikely, since
allocation should almost always succeed. Otherwise, we pass an
uninitialized string and length to `strtod`.

After seeing a suspicious test failure in the RelWithDebInfo build on
CI, I ran locally with Valgrind.