Guard for dict globals in function.__new__
`function.__new__` needs globals to be a dict (or, for us, a module
proxy would work too). The function was written in D20167616 and even
had a test for `function.__new__` with a non-dict globals, but this
relied on undefined behavior inside the runtime: frames used to be
`memset` to help `LOAD_FAST` check if a value was set, but this was
showing up in benchmarks. So instead we relaxed the constraint of
raising an `UnboundLocalError` and moved on with our lives. Now that we
are finally fixing this, we came across the bug of a conditionally
defined `mod`.

Finally, finally, finally, fix the bug! :)