Try not to raise AttributeError in hasattr
If we can avoid it, don't fully `strFormat` a string and set an
`AttributeError` on the `Thread`. There's no need; we're immediately
going to clear it.

To signal that we should not raise an `AttributeError`, set the initial
`location_out` to `Unbound::object()` (instead of the normal call path,
which sets it to `NoneType::object()`).