Use std::memmove in MutableTuple::replaceFromWithStartAt
This gets a lot of platform-specific optimizations and is a compiler
intrinsic. It should be able to handle overlap cases.

We may need to replace this later when we have read/write barriers, but
I think the idea of these functions is that they are supposed to be
atomic copies anyway.